switching to angular routes
