+++ sarf-ui/src/app/app.component.html
@@ -9,9 +9,6 @@
   </div>
 </mat-toolbar>
 <main>
-  <router-outlet>
-    <div>
-      <app-rootsearch></app-rootsearch>  
-    </div>
-  </router-outlet>  
+  <app-rootsearch></app-rootsearch>
+  <router-outlet></router-outlet>
 </main>
