+++ sarf-ui/src/app/start/start.component.html
@@ -1,0 +1,1 @@
+<p>أدخل جذر كلمة عربية</p>
