+++ sarf-ui/src/app/trilateral-conjugation-panel/trilateral-conjugation-panel.component.html
@@ -1,4 +1,4 @@
-<app-conjugation-group [conjugationGroup]="conjugationGroup.unaugmented"></app-conjugation-group>
-<app-conjugation-group [conjugationGroup]="conjugationGroup.augmentedByOne"></app-conjugation-group>
-<app-conjugation-group [conjugationGroup]="conjugationGroup.augmentedByTwo"></app-conjugation-group>
-<app-conjugation-group [conjugationGroup]="conjugationGroup.augmentedByThree"></app-conjugation-group>
+<app-conjugation-group [conjugationGroup]="conjugationGroup?.unaugmented"></app-conjugation-group>
+<app-conjugation-group [conjugationGroup]="conjugationGroup?.augmentedByOne"></app-conjugation-group>
+<app-conjugation-group [conjugationGroup]="conjugationGroup?.augmentedByTwo"></app-conjugation-group>
+<app-conjugation-group [conjugationGroup]="conjugationGroup?.augmentedByThree"></app-conjugation-group>
